+++ src/main/java/org/bot/nullbot/command/ai/embedding/RememberCommand.java
@@ -47,15 +47,15 @@ public class RememberCommand implements Command
             if (!sysMsgStorage.addLongTermUserMemory(targetId, params.getFirst()))
                 throw new NullBotMsgException("[记忆] ❌容量已满");
             bot.sendPrivateMsg(targetId, """
-                    [记忆] \uD83D\uDCA1 长时记忆已添加
-                    - 记忆: %s""".formatted(params.getFirst()), false);
+                    [记忆] \uD83D\uDCA1长时记忆已添加
+                    - 内容: %s""".formatted(params.getFirst()), false);
             log.info("\t\t\t\t├─[Remember] 用户长时记忆已添加 - {} : {}", targetId, params.getFirst());
         } else {
             if (!sysMsgStorage.addLongTermGroupMemory(targetId, params.getFirst()))
                 throw new NullBotMsgException("[记忆] ❌容量已满");
             bot.sendGroupMsg(targetId, """
-                    [记忆] \uD83D\uDCA1 长时记忆已添加
-                    - 记忆: %s""".formatted(params.getFirst()), false);
+                    [记忆] \uD83D\uDCA1长时记忆已添加
+                    - 内容: %s""".formatted(params.getFirst()), false);
             log.info("\t\t\t\t├─[Remember] 群聊长时记忆已添加 - {} : {}", targetId, params.getFirst());
         }
     }